pyproject.toml
[tool.poetry]
name = "pycollections"
version = "0.1.1"
description = ""
authors = ["jiangzihan <jiangzihan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-sdk = "^1.27.0"
opentelemetry-api = "^1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.27.0"
opentelemetry-exporter-otlp-proto-http = "^1.27.0"
opentelemetry-instrumentation-logging = "^0.48b0"
opentelemetry-exporter-prometheus = "^0.48b0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
